# docs/swagger.yaml
definitions:
...
  web.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  website.DelWebsiteRequest:
    properties:
      id:
        type: integer
    type: object
...
  website.SaveWebsiteRequest:
    properties:
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      link:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /website/delWebsite:
    delete:
      consumes:
      - application/json
      description: 删除网站信息
      parameters:
      - description: 删除网站信息请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/website.DelWebsiteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/web.Response'
      summary: 删除网站信息
      tags:
      - 网站
...
  /website/saveWebsite:
    post:
      consumes:
      - application/json
      description: 保存网站信息,id是可选字段,如果有就是替换原来的列表里的,如果没有就是存储新的值
      parameters:
      - description: 保存网站信息请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/website.SaveWebsiteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/web.Response'
      summary: 保存网站信息
      tags:
      - 网站
swagger: "2.0"